Move classes out of IoT Hub device client's "common" folder
These classes are moved to the most appropriate spot based on what set of folders contained classes that referenced them. Some classes were deleted because they were unused
