+++ iothub/device/src/Authentication/Security/ISharedAccessSignatureCredential.cs
