+++ iothub/device/src/Authentication/Security/SharedAccessSignature.cs
