+++ iothub/device/src/Authentication/Security/SharedAccessSignatureBuilder.cs
