+++ iothub/device/src/Authentication/Security/SharedAccessSignatureConstants.cs
