+++ iothub/device/src/Authentication/Security/StringValidationHelper.cs
