--- iothub/device/src/Common/WaitCallbackActionItem.cs
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.Azure.Devices.Client
-{
-    internal static class WaitCallbackActionItem
-    {
-        internal static bool ShouldUseActivity { get; set; }
-    }
-}
