+++ iothub/device/src/Exceptions/DeviceMaximumQueueDepthExceededException.cs
