+++ iothub/device/src/Exceptions/DeviceMessageLockLostException.cs
