+++ iothub/device/src/Exceptions/DeviceNotFoundException.cs
