+++ iothub/device/src/Exceptions/ExceptionHandlingHelper.cs
