+++ iothub/device/src/Exceptions/FatalException.cs
