+++ iothub/device/src/Exceptions/IotHubCommunicationException.cs
