+++ iothub/device/src/Exceptions/IotHubException.cs
