+++ iothub/device/src/Exceptions/IotHubSuspendedException.cs
