+++ iothub/device/src/Exceptions/QuotaExceededException.cs
