+++ iothub/device/src/Exceptions/ServerErrorException.cs
