+++ iothub/device/src/Transport/Amqp/AccessRights.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace Microsoft.Azure.Devices.Client
+namespace Microsoft.Azure.Devices.Client.Transport.Amqp
 {
     [Flags]
     [JsonConverter(typeof(StringEnumConverter))]
