+++ iothub/device/src/Transport/AmqpIot/SerializationUtilities.cs
