+++ iothub/device/src/Transport/ClientWebSocketTransport.cs
