+++ iothub/device/src/Transport/Mqtt/UrlEncodedDictionarySerializer.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Microsoft.Azure.Devices.Client.Common
+namespace Microsoft.Azure.Devices.Client.Transport.Mqtt
 {
     /// <summary>
     /// URL encoded serializer for message properties.
