+++ iothub/device/src/Utilities/ClientApiVersionHelper.cs
