+++ iothub/device/src/Utilities/CommonConstants.cs
