+++ iothub/device/src/Utilities/CommonExtensions.cs
