+++ iothub/device/src/Utilities/ExceptionExtensions.cs
