+++ iothub/device/src/Utilities/TaskHelpers.cs
