+++ iothub/device/src/Utilities/TimeoutHelper.cs
